+++ finagle-core/src/test/scala/com/twitter/finagle/util/StackRegistryTest.scala
@@ -5,33 +5,6 @@ import com.twitter.util.Var
 import com.twitter.util.registry.{Entry, GlobalRegistry, SimpleRegistry}
 import org.scalatest.FunSuite
 
-case class TestParam(p1: Int) {
-  def mk() = (this, TestParam.param)
-}
-object TestParam {
-  implicit val param = Stack.Param(TestParam(1))
-}
-
-case class TestParam2(p2: Int) {
-  def mk() = (this, TestParam2.param)
-}
-object TestParam2 {
-  implicit val param = Stack.Param(TestParam2(1))
-}
-
-class NotCaseClassParam(val ncc: Var[Int]) {
-  def mk() = (this, NotCaseClassParam.param)
-
-}
-object NotCaseClassParam {
-
-  implicit val param = new Stack.Param[NotCaseClassParam] {
-    val default = new NotCaseClassParam(Var(3))
-    override def show(p: NotCaseClassParam): Seq[(String, () => String)] =
-      Seq(("ncc", () => p.ncc.sample().toString))
-  }
-}
-
 class StackRegistryTest extends FunSuite {
 
   val headRole = Stack.Role("head")